% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pca.R
\name{add_pca}
\alias{add_pca}
\title{add pca variables to a data frame}
\usage{
add_pca(.data, ..., new_column = NULL, n = NULL)
}
\arguments{
\item{.data}{A data frame}

\item{...}{Comma separated list of unquoted expressions. You can treat 
variable names like they are positions. Use positive values to select 
variables; use negative values to drop variables.}

\item{new_column}{the initial name of the new columns to append. If left null
then variables will just be returned with .pc and the number for the
loading.}

\item{n}{the number of princple components to return. If left null then all 
loadings will be returned.}
}
\value{
the original data frame with the new principle component columns
}
\description{
\code{add_pca} takes a data frame and the columns that you wish to compute 
the pca values for. You can use \code{dplyr}'s select rules to pick the
columns you want or omit the ones you don't want to include. Only numeric
columns are allowed and any rows with missing values will be returned with
missing values for the PCA columns. It will append these new pca loadings to
the data frame as additional columns.
}
\examples{
add_pca(mtcars, mpg:wt, new_column = "car_specs", n = 3) \%>\% head(3)
}

