% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_model_evaluation.R
\name{binary_model_evaluation}
\alias{binary_model_evaluation}
\title{Calculate Binary Model Performance Metrics}
\usage{
binary_model_evaluation(.data, prediction_formula, group_var = NULL)
}
\arguments{
\item{.data}{a data frame}

\item{prediction_formula}{a formula specifiying the dependent binary variable
(lhs) and the probability predictions (rhs)}

\item{group_var}{an unquoted variable name to calculate performance measures
on each subset of the data - e.g., training/testing splits}
}
\value{
a data frame containing the metrics
}
\description{
Accepts a data frame containing the model predictions and actual values. The
following classification variables are calculated on the predictions:
}
\details{
\itemize{
  \item AUC - Area Under the Receiver Operating Curve 
  \item KS - Kolmogorov-Smirnov statistic
  \item MSE - Mean Square Error
  \item TPR - True Positive Rate
  \item TNR - True Negative Rate
  \item LSR - Logistic Scoring Rule (Log Loss)
  \item Bias - Direction of Bias
}
}
\examples{
iris_ <- iris
iris_$setosa <- ifelse(iris_$Species == "setosa", 1, 0)
iris_$pred <- iris_$Sepal.Length
iris_$pred <- (max(iris_$pred) - iris_$pred) / (max(iris_$pred) - min(iris_$pred) + 1)
iris_$splits <- sample(c("Test", "Train"), size = nrow(iris_), replace = T)
binary_model_evaluation(iris_, setosa ~ pred)
binary_model_evaluation(iris_, setosa ~ pred, group_var = splits)
}

