% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryModelValidation.R
\name{BinaryModelValidation}
\alias{BinaryModelValidation}
\title{Binary Model Validation}
\usage{
BinaryModelValidation(preds, Y, print_summary = T)
}
\arguments{
\item{preds}{a vector of values between 0 and 1 containing the predictions}

\item{Y}{the binary response variable.}

\item{print_summary}{should the KS, AUC, and RMSE be printed on the console, logical value.}
}
\value{
A list containing the AUC plot, Calibration plot, KS plot, Distribution plot, a plot containing all plots on one page, and a data frame containing the validation measures.
}
\description{
This function finds the KS Statistic, AUC, and RMSE value for a given 
predictor and binary response. In addition it returns the AUC, 
Calibration, KS, and Density plots for the predictor and response. The 
response vector can be a factor, character, or numeric vector. If Y is a 
character or factor then the highest level of the factor will be used as the
"1" level in a binary variable.
}
\examples{
set.seed(924516)
predictions <- runif(1000)
y_val <- 1 * (runif(1000) <= predictions + rnorm(1000, mean = 0, sd = .5))
val_plots <- BinaryModelValidation(preds = predictions, Y = y_val, print_summary = F)
}
\keyword{eda}

