% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{deciles}
\alias{deciles}
\title{returns the min, deciles, and max for a numeric vector}
\usage{
deciles(x)
}
\arguments{
\item{x}{a numeric or integer vector}
}
\value{
a numeric vector with names for the corresponding decile
}
\description{
\code{deciles} takes in a numeric vector and returns a named vector that
contains the min, 10, 20, 30, 40, 50, 60, 70, 80, and 90th percentiles,
and the max. Missing values are always removed
}
\examples{
deciles(0:100)
}

