% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_woe.R
\name{find_woe}
\alias{calc_woe}
\alias{find_woe}
\title{finds weight of evidence for independent variables}
\usage{
find_woe(.data, ..., y)

calc_woe(x, y, bins)
}
\arguments{
\item{.data}{a data frame containing the variables to analyze}

\item{...}{columns to calculate woe for. Uses \code{dplyr}'s \code{select} rules}

\item{y}{the dependent binary variable}
}
\value{
a data frame where each row lists a columns Information Value and a 
nested data frame containing the splits used and the Weight of Evidence for
each.
}
\description{
Measures and returns the Weight of Evidence (WOE) and Information Value (IV)
for selected independent variables against a specified binary dependent 
variable.
}
\section{Functions}{
\itemize{
\item \code{calc_woe}: calculates woe
}}
\examples{

find_woe(mtcars, mpg:qsec, y = am)
}

