% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{grouped_arrange}
\alias{grouped_arrange}
\title{arrange a data frame including the grouped columns}
\usage{
grouped_arrange(.data, ...)
}
\arguments{
\item{.data}{a data frame}

\item{...}{the columns to arrange. Uses \code{dplyr}'s \code{select} rules}
}
\description{
this function will automatically include the grouped columns in the arrange
call.
}

