% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_code.R
\name{impact_code}
\alias{impact_code}
\title{transform high cardinality categorical variables}
\usage{
impact_code(.data, formula, binary = T)
}
\arguments{
\item{.data}{the data frame the data is contained in}

\item{formula}{a formula specifying the dependent and independent variables to use}

\item{binary}{is the dependent variable a binary variable? Logical value}
}
\value{
a data frame that includes the levels of the categorical variable and the smoothed estimate of the dependent variable.
}
\description{
\code{impact_code} transforms a categorical variable into a continuous one 
using Impact Coding described in this
\href{http://helios.mm.di.uoa.gr/~rouvas/ssi/sigkdd/sigkdd.vol3.1/barreca.pdf}{paper}.
}
\details{
Missing values in the categorical variable will be treated as another level. 
If you want to ignore missing values you will need to do so prior to calling
\code{impact_code}.
}
\examples{
impact_code(mtcars, am ~ cyl)
impact_code(iris, Petal.Width ~ Species, binary = F)
}

