% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KS.r
\name{KS}
\alias{KS}
\title{Find KS Statistic of an Independent - Dependent Variable Combo}
\usage{
KS(x, y)
}
\arguments{
\item{y}{Must be Categorical or only have two values. Also can't contain any \code{NA} values}
}
\value{
The KS value for the independent variable
}
\description{
This function calculates the KS value associated with a linear model made
from a single predictor on a binary dependent variable. It also removes any
missing values from the dependent variable before calculating KS.
}
\examples{
KS(x = 1:100, y = sample(c(0, 1), 100, replace = T))
}
\keyword{ks}

