% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{model_matrix}
\alias{model_matrix}
\title{Preserves all factor levels when creating a model matrix}
\usage{
model_matrix(object, .data)
}
\arguments{
\item{object}{the formula for creating the model}

\item{.data}{a data frame}
}
\value{
a numeric matrix to be used for modeling purposes
}
\description{
In some cases it's easier to not have a reference level for dummy variables.
This function will preserve all levels of character and factor variables when
using model.matrix.
}
\examples{
head(model_matrix(iris, Petal.Width ~ Petal.Length + Species))
}

