% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{n_combos}
\alias{n_combos}
\title{n_distinct for data frames}
\usage{
n_combos(.data, ...)
}
\arguments{
\item{.data}{a data frame}

\item{...}{the columns to count the number of groups. Uses \code{dplyr}'s \code{select} rules}
}
\description{
returns the number of distinct combinations for the selected columns
}

