% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_plot.R
\name{partial_plot}
\alias{partial_plot}
\title{provides partial plots for fitted \code{gam} models}
\usage{
partial_plot(fitted_model, variable, response = F)
}
\arguments{
\item{fitted_model}{a complete \code{gam} model object}

\item{variable}{the name of the independent variable as a character string}

\item{response}{logical indicating if the plot should be on the linear 
prediction scale or the response scale. Defaults to \code{FALSE}}
}
\value{
a \code{ggplot2} object of the partial regression plot
}
\description{
\code{partial_plot} accepts a fitted \code{gam} object and the name of the
variable you wish to view the partial regression plot of as a character
string. It returns a \code{ggplot2} object with the values of the
independent variable plotted against the spline predictions plus the 
intercept of the model.
}
\examples{
# library(ISLR)
# library(mgcv)
# default_gam <- gam(default ~ s(balance, fx = T, k = 6), data = Default, family = "binomial") 
}

