% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_plot.R
\name{patrial_plot}
\alias{patrial_plot}
\title{provides partial plots for fitted \code{gam} models}
\usage{
patrial_plot(fitted_model, variable)
}
\arguments{
\item{fitted_model}{a complete \code{gam} model object}

\item{variable}{the name of the independent variable as a character string}
}
\value{
a \code{ggplot2} object of the partial regression plot
}
\description{
\code{partial_plot} accepts a fitted \code{gam} object and the name of the
variable you wish to view the partial regression plot of as a character
string. It returns a \code{ggplot2} object with the values of the
independent variable plotted against the spline predictions plus the 
intercept of the model.
}

