% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peruse.R
\name{peruse}
\alias{peruse}
\title{Peruse a Data Frame}
\usage{
peruse(df)
}
\arguments{
\item{data}{a data frame}
}
\description{
\code{peruse} examines the variables in a data frame and returns basic 
summary info about the individual variables. Different values will be 
returned for different types of columns so the function returns a data frame 
with the underlying data nested.
}
\examples{
cars_summary <- peruse(mtcars)
numeric_info <- cars_summary \%>\%
  dplyr::filter(Class == "Numeric") \%>\%
  tidyr::unnest(data)

}

