% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\alias{profile.character}
\alias{profile.factor}
\alias{profile.logical}
\alias{profile.numeric}
\title{Profile variables}
\usage{
profile(element)

\method{profile}{numeric}(element)

\method{profile}{character}(element)

\method{profile}{factor}(variable)

\method{profile}{logical}(variable)
}
\arguments{
\item{element}{a vector you wish to profile.}
}
\description{
\code{profile} takes in a vector and, depending on the class of the vector,
will return summary information on that vector returned as a data.frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: profile numeric vectors

\item \code{character}: profile character vectors

\item \code{factor}: profile factor vectors

\item \code{logical}: profile factor vectors
}}
\examples{
set.seed(123456)
sample_variable <- sample(letters, size = 100, replace = T)
profile(sample_variable)

}

