% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_from}
\alias{remove_from}
\alias{remove_na}
\alias{remove_null}
\title{Removes values from a vector}
\usage{
remove_from(vec, value)

remove_null(vec)

remove_na(vec)
}
\arguments{
\item{vec}{the vector to remove values from}

\item{value}{the value to remove. The value can either be a single value or
a vector of values}
}
\value{
the original vector with the values removed
}
\description{
The \code{remove_} functions are convenient wrappers to help remove specific
values from a vector. \code{remove_from} will remove single or multiple
values from a given vector. In addition the \code{remove_null} and
\code{remove_na} functions will remove the corresponding missing values.
}
\section{Functions}{
\itemize{
\item \code{remove_null}: Removes null values

\item \code{remove_na}: Removes NA values
}}

