% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_groups.R
\name{sample_groups}
\alias{sample_groups}
\title{sample from groups of variables}
\usage{
sample_groups(.data, ..., n = 6)
}
\arguments{
\item{...}{the variables you would like to group}

\item{n}{the number of groupings to return}

\item{df}{the data.frame you are sampling from}
}
\value{
a data frame containing the sampled groupings
}
\description{
\code{sample_groups} allows you to restrict the data to only a limited number
of groupings of variables. This is useful for when you have multilevel data
and would like to sample from multiple levels at once. This makes exploratory
plotting a little easier.
}
\examples{
sample_groups(mtcars, cyl, am)
}

