% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_y_percentage}
\alias{scale_x_percentage}
\alias{scale_y_percentage}
\title{use percentage scales for ggplot2}
\usage{
scale_y_percentage(...)

scale_x_percentage(...)
}
\description{
adds percentage scales to ggplot2 plots
}
\section{Functions}{
\itemize{
\item \code{scale_x_percentage}: adds percentage scales to ggplot2 plots
}}

