% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_code.R
\name{smoother_binomial}
\alias{smoother_binomial}
\alias{smoother_binomial.character}
\alias{smoother_binomial.factor}
\alias{smoother_binomial.logical}
\alias{smoother_binomial.numeric}
\title{helper functions for smoother_binomial}
\usage{
smoother_binomial(y, x, groups)

\method{smoother_binomial}{factor}(y, x, groups)

\method{smoother_binomial}{logical}(y, x, groups)

\method{smoother_binomial}{character}(y, x, groups)

\method{smoother_binomial}{numeric}(y, x, groups)
}
\description{
helper functions for smoother_binomial
}
\section{Methods (by class)}{
\itemize{
\item \code{factor}: factors

\item \code{logical}: logicals

\item \code{character}: characters

\item \code{numeric}: numerics
}}

