% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableNA.R
\name{tableNA}
\alias{tableNA}
\title{Uses the base \code{table} function but always checks for NAs}
\usage{
tableNA(...)
}
\arguments{
\item{...}{the object you want to use the \code{table} function on}
}
\value{
a \code{table}
}
\description{
Uses the base \code{table} function but always checks for NAs
}
\examples{
car_copy <- mtcars
car_copy$cyl[c(2, 7, 15)] <- NA
table(car_copy$cyl)
tableNA(car_copy&cyl)
}

