% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/VariableExplore.r
\name{VariableExplore}
\alias{VariableExplore}
\title{Variable Exploration}
\usage{
VariableExplore(x, y, var_name, plot_save = F, fname = "none")
}
\arguments{
\item{x}{independent variable, must be a vector}

\item{y}{dependent variable, must be a vector and can't contain missing values}

\item{var_name}{The name of the dependent variable}

\item{plot_save}{logical indicating if plot should be saved, default is FALSE}

\item{fname}{the name of the file to save the pdf output}
}
\value{
A list containing a plot showing all the info as well as the information that makes it up
}
\description{
This function produces a histogram of the independent variable,
summary statics, missing data information, and estimates of the trend
between it and the dependent variable.
}
\examples{
VariableExplore(1:100, sample(c(0, 1), 100, replace = T))
}
\keyword{eda}

