% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{Prints samples from data structre}
\usage{
view(x, n = 6)
}
\arguments{
\item{x}{A data frame or vector}

\item{n}{the number of rows or values to print}
}
\value{
the printed values
}
\description{
\code{view} randomly samples from the input and prints the selected values
to the console. This helps see values that \code{head} and \code{tails} may
not reach.
}
\examples{
view(letters)
view(mtcars, n = 4)

}

