% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/WOE.r
\name{WOE}
\alias{WOE}
\title{Find Weight Of Evidence}
\usage{
WOE(x, y, bins = 10, adj = 0.5, incl_NA = T)
}
\arguments{
\item{x}{Independent variable}

\item{y}{Must be Categorical or only have two values. Also can't contain any \code{NA} values}

\item{bins}{The number of buckets to bin the independent variable (\code{x})}

\item{adj}{A small adjustment should be included if some buckets may only have 1 level of \code{y}. Defaults to .5}

\item{incl_NA}{should \code{NA} observations be included? Default = TRUE}
}
\value{
A list containing:

\code{InformationValue} The information value from the WOE

\code{Bins} The cutpoints for the bins

\code{Plot} ggplot object with the WOE of each bin.
}
\description{
This function finds the WOE and IV of a variable in regards to its relationship with the dependent variable
}
\examples{
result <- WOE(x = 1:100, y = sample(c(0, 1), 100, replace = T))
}
\keyword{IV}
\keyword{WOE,}

