% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_drift.R
\name{check_drift}
\alias{check_drift}
\title{Visualize trend in variables across another}
\usage{
check_drift(.data, formula, bins = NULL)
}
\arguments{
\item{.data}{a data frame containing the data in question}

\item{formula}{a formula with the variable on the lhs and grouping variable 
on the rhs}

\item{bins}{the number of bins to make. If \code{NULL} (default) then each 
variable value will be used.}
}
\value{
a ggplot2 object showing the distribution of the variable colored by
group
}
\description{
Returns a plot showing the distribution of a variable by different levels of
another grouping variable. A common application in my work is making sure the
distribution doesn't change across different seasons.
}
\examples{
check_drift(iris, Petal.Width ~ Species, bins = 5)
check_drift(mtcars, cyl ~ am, bins = NULL)
}

