% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/MissingImputation.r
\name{MissingImputation}
\alias{MissingImputation}
\title{Missing Imputation}
\usage{
MissingImputation(missing_df, num_iter = 10)
}
\arguments{
\item{missing_df}{A data frame containing at least some columns with missing values}

\item{num_iter}{Number of iterations to perform}
}
\value{
A list containing
\code{complete_obs} a data frame with the missing values replaced through regression imputation
\code{change} The differences between each iteration for each variable
}
\description{
This function completes regression-based missing imputation. The function accepts a data frame containing variables with missing values. The data frame can contain variables with no missing values if you would not like to include them in replacing the values with missing variables.
}
\examples{
iris_df <- iris
set.seed(123456)
na1 <- runif(nrow(iris_df)) < .25
na2 <- runif(nrow(iris_df)) < .33
iris_df[["Species"]][na1] <- NA
iris_df[["Sepal.Length"]][na2] <- NA
iris_complete <- MissingImputation(iris_df, num_iter = 5)
}
\keyword{eda}

