% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_woe.R
\name{woe}
\alias{woe}
\alias{woe.character}
\alias{woe.factor}
\alias{woe.numeric}
\title{Find Weight Of Evidence}
\usage{
woe(x, y)

\method{woe}{numeric}(x, y)

\method{woe}{factor}(x, y)

\method{woe}{character}(x, y)
}
\arguments{
\item{x}{Independent variable}

\item{y}{Must be Categorical or only have two values. Also can't contain any \code{NA} values}
}
\description{
This function finds the WOE and IV of a variable in regards to its 
relationship with the dependent variable. If x is a factor it will be 
converted to a character vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: numeric

\item \code{factor}: factor

\item \code{character}: character
}}
\keyword{IV}
\keyword{WOE,}

